requirements.txt
gdown==5.2.0
pytorch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
cpuonly==2.0